DataType: Cleanup
