+++ core/src/main/scala/org/scalarelational/datatype/SimpleDataType.scala
@@ -1,0 +1,10 @@
+package org.scalarelational.datatype
+
+/**
+ * @author Matt Hicks <matt@outr.com>
+ */
+class SimpleDataType[T](jdbcType: Int,
+                        sqlType: SQLType,
+                        converter: SQLConversion[T, T] = SQLConversion.identity[T],
+                        sqlOperator: SQLOperator[T, T] = new DefaultSQLOperator[T, T])
+  extends DataType[T, T](jdbcType, sqlType, converter, sqlOperator)
